# avoid allocations when unmarshaling