--- a/resp/resp.go
+++ b/resp/resp.go
@@ -78,6 +78,22 @@ func putBytes(b *[]byte) {
 	bytePool.Put(b)
 }
 
+var bufferPool = sync.Pool{
+	New: func() interface{} {
+		b := make([]byte, 0, bytes.MinRead)
+		return bytes.NewBuffer(b)
+	},
+}
+
+func getBuffer() *bytes.Buffer {
+	return bufferPool.Get().(*bytes.Buffer)
+}
+
+func putBuffer(b *bytes.Buffer) {
+	b.Reset()
+	bufferPool.Put(b)
+}
+
 // parseInt is a specialized version of strconv.ParseInt that parses a base-10 encoded signed integer.
 func parseInt(b []byte) (int64, error) {
 	if len(b) == 0 {
@@ -179,12 +195,17 @@ func bufferedIntDelim(br *bufio.Reader) (int64, error) {
 }
 
 func readAllAppend(r io.Reader, b []byte) ([]byte, error) {
-	buf := bytes.NewBuffer(b)
+	buf := getBuffer()
 	// a side effect of this is that the given b will be re-allocated if
 	// it's less than bytes.MinRead. Since this b could be all the way from the
 	// user we can't guarantee it within the library
 	_, err := buf.ReadFrom(r)
-	return buf.Bytes(), err
+	b = append(b, buf.Bytes()...)
+	putBuffer(buf)
+	if b == nil {
+		b = []byte{}
+	}
+	return b, err
 }
 
 func multiWrite(w io.Writer, bb ...[]byte) error {