Currently `readAllAppend` allocates a `*bytes.Buffer` on each call. Also for slices with less than bytes.MinRead bytes (512) the buffer will allocate a new `[]byte` even if the old slice could fit all the data.

In most cases the capacity of the `[]byte` is rather small (64 bytes in most cases, which is the capacity used by `getBytes` for new slices), which means that most calls will allocate at least 3 times.

Avoid this by reading into pooled buffers and appending the data to the given `[]byte`.

Also add a benchmark that calls `readAllAppend` with different capacities to show the different allocation cases.

Benchmark results:

```
name                         old time/op    new time/op    delta
ReadAllAppend/Capacity0-8       965ns ± 3%      99ns ± 3%   -89.76%  (p=0.000 n=15+15)
ReadAllAppend/Capacity11-8      439ns ± 5%      66ns ± 2%   -85.02%  (p=0.000 n=15+13)
ReadAllAppend/Capacity523-8     121ns ± 6%      67ns ± 3%   -44.45%  (p=0.000 n=15+15)

name                         old alloc/op   new alloc/op   delta
ReadAllAppend/Capacity0-8      2.16kB ± 0%    0.02kB ± 0%   -99.26%  (p=0.000 n=15+15)
ReadAllAppend/Capacity11-8       688B ± 0%        0B       -100.00%  (p=0.000 n=15+15)
ReadAllAppend/Capacity523-8      112B ± 0%        0B       -100.00%  (p=0.000 n=15+15)

name                         old allocs/op  new allocs/op  delta
ReadAllAppend/Capacity0-8        3.00 ± 0%      1.00 ± 0%   -66.67%  (p=0.000 n=15+15)
ReadAllAppend/Capacity11-8       2.00 ± 0%      0.00       -100.00%  (p=0.000 n=15+15)
ReadAllAppend/Capacity523-8      1.00 ± 0%      0.00       -100.00%  (p=0.000 n=15+15)
```