@nussjustin 2018-05-19
More context on this: I have a program that reads a set of hash keys and calls HGETALL for each key in a pipeline. Avoiding the allocations in readAllAppend improves performance by at least 10% even for small numbers of keys.

Benchmarks for different set sizes:

```
name            old time/op    new time/op    delta
ReadAll/1-8        123µs ± 3%     117µs ± 5%     ~     (p=0.056 n=5+5)
ReadAll/10-8       507µs ± 3%     452µs ± 4%  -10.88%  (p=0.008 n=5+5)
ReadAll/100-8     4.88ms ± 3%    3.76ms ± 1%  -22.92%  (p=0.008 n=5+5)
ReadAll/1000-8    50.7ms ± 1%    44.7ms ± 2%  -11.94%  (p=0.008 n=5+5)

name            old alloc/op   new alloc/op   delta
ReadAll/1-8       11.5kB ± 0%     1.2kB ± 0%  -89.43%  (p=0.000 n=5+4)
ReadAll/10-8       122kB ± 0%      19kB ± 0%  -84.21%  (p=0.008 n=5+5)
ReadAll/100-8     1.22MB ± 0%    0.18MB ± 0%  -84.95%  (p=0.008 n=5+5)
ReadAll/1000-8    12.3MB ± 0%     1.9MB ± 0%  -84.21%  (p=0.008 n=5+5)

name            old allocs/op  new allocs/op  delta
ReadAll/1-8          161 ± 0%        69 ± 0%  -57.14%  (p=0.008 n=5+5)
ReadAll/10-8       1.62k ± 0%     0.70k ± 0%  -56.86%  (p=0.008 n=5+5)
ReadAll/100-8      16.2k ± 0%      7.0k ± 0%  -56.78%  (p=0.008 n=5+5)
ReadAll/1000-8      162k ± 0%       70k ± 0%  -56.64%  (p=0.008 n=5+5)
``` 

The benchmark only include the HGETALL commands not getting the keys.